add storage for IP and wifi settings
To change the wifi settings in a running system there is an storage area in the config.
If there is soemthing stroed, the WifiManager tries to connect to the stored ssid.

To get the possibility to st a static IP address, this is now stored in the config.
As deflault there is set DHCP, to make no change in default configuration.